multinode tests: Add ping tests for vif and container migration.
For the migration scenario, we were testing the migration of a port p1
from src to dst, and checking whether ping to/from a port p2 in a
3rd node was fine.
We now added testing ping from p1 to ports in src, dst and 3rd node.

The IPs have been changed to make them more 'logical' i.e. simplify debugging.

Signed-off-by: Xavier Simonart <xsimonar@redhat.com>