tests: Fix flaky "tag allocation".
When a container port is added to a ls, with a new tag allocation being
requested (e.g. ovn-nbctl lsp-add ls1 c1 parent1 0), this is done by
northd in two iterations: the sb pb is created, and then the tag is allocated.
Hence, there is no guarantee that the tag is allocated in sb when
ovn-nbctl --wait=sb returns.
Before [1], this was not causing any issue, as the second transaction
(writing the tag in the sb db) was initiated before ovn-nbctl --wait=sb
returned. But this is not the case anymore.

[1] e1495fbc86bb: "northd: Remove delay in sb_cfg propagation.".

Signed-off-by: Xavier Simonart <xsimonar@redhat.com>